Update rgb_matrix_get_mode to return uint8_t.

+++ quantum/rgb_matrix.h
@@ -180,7 +180,7 @@ void rgb_matrix_increase_speed(void);
 void rgb_matrix_decrease_speed(void);
 void rgb_matrix_mode(uint8_t mode);
 void rgb_matrix_mode_noeeprom(uint8_t mode);
-uint32_t rgb_matrix_get_mode(void);
+uint8_t rgb_matrix_get_mode(void);
 
 #ifndef RGBLIGHT_ENABLE
 #define rgblight_toggle() rgb_matrix_toggle()
